1.세션값 PostMapping에선 못받아오는 문제점 해결하기(test성공) 2.파일 업로드/ 다운로드 3.파일 다운로드할때 save이름이 아닌 original이름으로 바꾸기(디비에서 꺼내옴) 4.소리파일 사운드출력 구현(.wav파일만 가능) 5.board 테이블 no, seq_no 값 설정(디비에 No칼럼값을 Big int로 바꿈) -게시판 글 삭제할때마다 삭제한것보다 하위일련번호가 높은것들은 하위일련번호가 1씩 감소하는것 구현

+++ Spring_Dictation/Dictation/src/main/java/com/dictation/vo/BoardVO.java
@@ -12,7 +12,7 @@ public class BoardVO {
 	private String term;
 	private int lecture_no;
 	private String board_cd;
-	private int no;
+	private long no;
 	private int seq_no;
 	private String title;
 	private String content;
